+++ dwds/test/instances/patterns_inspection_test.dart
@@ -15,7 +15,7 @@ import 'common/patterns_inspection_common.dart';
 
 void main() {
   // Enable verbose logging for debugging.
-  final debug = false;
+  const debug = false;
   final canaryFeatures = false;
 
   group('canary: $canaryFeatures |', () {