+++ webdev/test/e2e_test.dart
@@ -38,7 +38,7 @@ final _testItems = <String, bool?>{
 
 void main() {
   // Change to true for debugging.
-  final debug = false;
+  const debug = false;
 
   final testRunner = TestRunner();
   late String soundExampleDirectory;